now calling shop product to display items in cart
