--- gracefabrics/src/components/Basket/basket.css
@@ -1,3 +1,0 @@
-#basketImg{
-    max-width: 120px;
-}
